fixing the connection to database for reservation and the how the reservation should format with the time and move evrything to componants and make page looks tidy
