+++ app/restaurant-details.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { StyleSheet, Text, View } from "react-native";
+export default function RestaurantDetails() {
+  return (
+    <View>
+      <Text></Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({});
